+++ example/gilded_rose/lib/gilded_rose.dart
@@ -20,7 +20,7 @@ final class GildedRose {
           }
         }
       } else {
-        if (items[i].quality < 550) {
+        if (items[i].quality < 50) {
           items[i].quality = items[i].quality + 1;
 
           if (items[i].name == "Backstage passes to a TAFKAL80ETC concert") {