+++ example/verify_methods/verify/verify_test.dart
@@ -3,8 +3,8 @@ import 'package:test/test.dart';
 
 void main() {
   test('Verify method', () {
-    // Simulated response from a function
-    String response = '{"result": "success", "data": {"id": 1, "name": "Item"}}';
+    String response =
+        '{"result": "success", "data": {"id": 1, "name": "Item"}}';
 
     try {
       ApprovalTests.verify(