+++ example/verify_methods/verify_all_combinations/verify_all_combinations_test.dart
@@ -1,0 +1,16 @@
+import 'package:approval_tests/approval_dart.dart';
+import 'package:test/test.dart';
+
+void main() {
+  test('Verify all combinations of input values', () {
+    List<List<int>> inputs = [
+      [1, 2],
+      [3, 4]
+    ];
+
+    ApprovalTests.verifyAllCombinations(
+      inputs: inputs,
+      processor: (combination) => 'Combination: ${combination.join(", ")}',
+    );
+  });
+}