+++ example/verify_methods/verify_as_json/json_test.dart
@@ -9,7 +9,9 @@ void main() {
         'features': ['Testing', 'JSON'],
         'version': 0.1,
       };
-      ApprovalTests.verifyAsJson(complexObject);
+      ApprovalTests.verifyAsJson(
+        complexObject,
+      );
     });
   });
 }